refactor: reclocating auth dto file and fixing minor bugs on auth controller
